refactor(voice): share OpenAI config/call-state base classes for Media Streams providers
Extracts MediaStreamsOpenAIProviderConfig and MediaStreamsOpenAICallState so
OpenAIRealtimeProviderConfig/_CallState inherit shared fields instead of
duplicating them, and adds MediaStreamsOpenAIProvider as shared scaffolding
for a Media Streams provider bridging to an OpenAI real-time voice API.

Also trims module/class docstrings that named specific subclasses or counted
"the only implementation today" — that phrasing goes stale the moment a new
subclass is added and creates comment upkeep with no payoff.

This is prep work so a future Media Streams provider only needs to add its
own package under media_streams/, without touching these shared files.
