+++ src/tac/channels/voice/media_streams/__init__.py
@@ -1,3 +1,3 @@
 """Voice providers bridging Twilio Media Streams (``<Connect><Stream>``) to a
-speech-to-speech model. ``openai_realtime`` is the first implementation.
+speech-to-speech model.
 """
