+++ src/tac/channels/voice/media_streams/shared/__init__.py
@@ -1,0 +1,4 @@
+"""Code shared across Media Streams providers that doesn't belong to any one
+of them — e.g. ``MediaStreamsOpenAIProvider``, the base for providers
+bridging to an OpenAI real-time voice API.
+"""
